% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{print.CDM}
\alias{print.CDM}
\title{Print Method for CDM Objects}
\usage{
\method{print}{CDM}(x, ...)
}
\arguments{
\item{x}{A \code{\link[Qval]{CDM}} object to be printed.}

\item{...}{Additional arguments.}
}
\description{
This function prints the details of a \code{\link[Qval]{CDM}} object.
It outputs the call used to create the object, the version and the date of the \code{Qval} package.
}
