% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{print.sim.data}
\alias{print.sim.data}
\title{Print Method for sim.data Objects}
\usage{
\method{print}{sim.data}(x, ...)
}
\arguments{
\item{x}{A \code{\link[Qval]{sim.data}} object to be printed.}

\item{...}{Additional arguments.}
}
\description{
This function prints the details of a \code{\link[Qval]{sim.data}} object.
It outputs the call used to create the object, the version and the date of the \code{Qval} package.
}
