% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q_val.R
\name{q_val}
\alias{q_val}
\title{Calculate FDR-adjusted q-values from a vector of p-values}
\usage{
q_val(p_vals, q_res = 0.001, sharpened = TRUE)
}
\arguments{
\item{p_vals}{Numerical vector with the p values to calculate from}

\item{q_res}{Defaults to 0.001, the resolution of the resulting q value}

\item{sharpened}{Logical, if TRUE then uses sharpened q values calculation}
}
\value{

}
\description{
Calculate FDR-adjusted q-values from a vector of p-values
}
