% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mmatrix.R
\name{get.Mmatrix}
\alias{get.Mmatrix}
\title{Calculate \eqn{\mathbf{M}} matrix}
\usage{
get.Mmatrix(K = NULL, pattern = NULL)
}
\arguments{
\item{K}{The number of attributes. Can be NULL if \code{pattern} is passed to the function and is not NULL.}

\item{pattern}{The knowledge state matrix containing all possible attribute mastery pattern.
Can be gained from @seealso \code{\link[GDINA]{attributepattern}}. Also can be NULL if \code{K}
is passed to the function and is not NULL.}
}
\value{
An object of class \code{matrix}.
}
\description{
Calculate \eqn{\mathbf{M}} matrix for stauted CDMs (de la Torre, 2011).
}
\examples{

library(Qval)

example.Mmatrix <-  get.Mmatrix(K = 5)

}
\references{
de la Torre, J. (2011). The Generalized DINA Model Framework. Psychometrika, 76(2), 179-199. https://doi.org/10.1007/s11336-011-9207-7.
}
\author{
Haijiang Qin <Haijiang133@outlook.com>
}
